--- a/export_mapof_test.go
+++ b/export_mapof_test.go
@@ -3,6 +3,6 @@
 
 package xsync
 
-func CollectMapOfStats[V any](m *MapOf[V]) MapStats {
+func CollectMapOfStats[K, V any](m *MapOf[K, V]) MapStats {
 	return MapStats{m.stats()}
 }