@iamcalledrob 2022-08-13
> Thanks for the contribution. I'd like to address the second consideration from [#17 (comment)](https://github.com/puzpuzpuz/xsync/issues/17#issuecomment-1213339363) before merging this one. I'm going to address it, but probably not in the nearest term. Will post an update here once it's done.

No worries, just throwing this up for discussion. No pressure to merge :)