@iamcalledrob 2022-08-13
It's worth noting that I'm seeing VERY poor performance for certain keys. This could be an issue I've introduced here, or it could be that the string hashes were just unlikely to generate problematic keys.

Here's an example that's very problematic on my machine (Apple M1 Pro):

```go
// Benchmark32768000000/xsync.Map-8         	     142	  20367105 ns/op
// Benchmark32768000000/sync.Map-8          	 4306820	       385.9 ns/op
func Benchmark32768000000(b *testing.B) {
	b.Run("xsync.Map", func(b *testing.B) {
		m := NewIntegerMapOf[uint64, int]()
		for i := 0; i < b.N; i++ {
			m.Store(uint64(i*32768000000), 0)
		}
	})
	b.Run("sync.Map", func(b *testing.B) {
		m := sync.Map{}
		for i := 0; i < b.N; i++ {
			m.Store(uint64(i*32768000000), 0)
		}
	})
}

// Benchmark32768000001/xsync.Map-8         	21440982	        48.54 ns/op
// Benchmark32768000001/sync.Map-8          	 4475781	       385.2 ns/op
func Benchmark32768000001(b *testing.B) {
	b.Run("xsync.Map", func(b *testing.B) {
		m := NewIntegerMapOf[uint64, int]()
		for i := 0; i < b.N; i++ {
			m.Store(uint64(i*32768000001), 0)
		}
	})
	b.Run("sync.Map", func(b *testing.B) {
		m := sync.Map{}
		for i := 0; i < b.N; i++ {
			m.Store(uint64(i*32768000001), 0)
		}
	})
}
```