@puzpuzpuz 2022-10-15
> I'd like to address the second consideration from [#17 (comment)](https://github.com/puzpuzpuz/xsync/issues/17#issuecomment-1213339363) before merging this one. I'm going to address it, but probably not in the nearest term. Will post an update here once it's done.

@iamcalledrob I've just merged #48, so this PR should be unblocked at last. Sorry for taking that long. :smiling_face_with_tear: 

Let me know if you're willing to give another try with this PR, otherwise I'll pick it up myself.