@puzpuzpuz 2022-10-15
> It's worth noting that I'm seeing VERY poor performance for certain keys.

This is very likely due to hash collisions and #48 should make this problem less significant.

The problem with `32768000000 * i` values is that all of them will have lots of zeros in LSBs (least significant bits), so they were ending up in the same bucket leading to lots of map resize operations.