@puzpuzpuz 2022-10-22
Closing this one in favor of #49 to speed up things.

@iamcalledrob many thanks for your contribution!